[7.2.0] feat(cc toolchain): support external_include_paths on macos (#22413)
The external_include_paths feature enables specifying include paths
locally with -I and with -isystem when the project is included as an
external repo.

This makes it possible to set -Werror against your own headers for
development without propagating this to consumers of your libraries.

For some reason this was only previously enabled on linux. I do not see
a reason for not also enabling it on macos.

Closes #21949.

PiperOrigin-RevId: 634394452
Change-Id: I0df91b2a2c9b4a5bd52b22e64f99ea0745cb9759

Commit
https://github.com/bazelbuild/bazel/commit/d831214a6d3ed7ae6b4c14ce4735f023b1da6e3c

Co-authored-by: Isaac Torres <isaac@isaactorz.dev>