Ignore the main repository when validating inclusions in a repository whose name is the same as the main repository.
This sounds like a weird corner case, but it happens in one very prominent case, which is protobuf: it contains a `local_repository(path=".")` rule in its WORKSPACE file in order to work around the consequences of the uncomfortable fact that `@bazel_tools` depends on it.

If this was *not* done, when sibling repository layout is in effect, paths in the external repo would be resolved as ones in the main repo, thus causing "undeclared inclusion" errors when running C++ actions in the external repo.

This is not perfect (after all, it's ambiguous which Artifact such paths refer to), but seems to be good enough for protobuf and will go away once the `@bazel_tools -> @com_google_protobuf` dependency is cut or if bzlmod is turned on for good.

RELNOTES: None.
PiperOrigin-RevId: 634663132
Change-Id: Id45eaa9bdd7746b33166e838dbb834febbec9335