Make it possible to dump Skyframe memory at the end of a build.
It works by passing the `--experimental_skyframe_memory_dump=json` command line option to build-like commands. If `--experimental_stream_log_file_uploads` is in effect, the file will be streamed to remote storage, if not, it'll be written to the output base to the eponymous file. Either way, its URI will be reported on the BEP as the file `skyframe_memory.json`.

The format is the same as `bazel dump --memory=full,summary` and befitting an experimental option, is subject to change.

This is modeled after the now-defunct `--experimental_aquery_dump_after_build_format` (that one will be deleted after this change is submitted)

RELNOTES: None.
PiperOrigin-RevId: 634743262
Change-Id: I3b5b4e13a252c7e6f7b4bad3067ea88753ff5931