Disable absl debug logging when `--client_debug=false` (the default).
The Bazel client codebase doesn't use absl, but one of the library we use (e.g. gRPC) might. We wouldn't want that to spam the Bazel's user's stderr. Thus, this CL attempts to insulate the client program from that risk.

We do this by disabling absl debug logging at the point in the life of the client program where we consider `--client_debug`. This is therefore a non-foolproof approach; consider a library we use that makes debug statements before that point in the life of our program. But we also add an integration test that would catch regressions like this!

I noted future improvements in code comments.

PiperOrigin-RevId: 635488965
Change-Id: I26d01f9aeda3a237aa68438586f958ae01d5448d