In 'test'/'run' commands, BEP will extract "non-crash" OOM error details to set `AbortReason`.
When the main command thread is interrupted during a build, it may be due to a
catastrophic error that happened on another thread. In this scenario we already
correctly capture the `DetailedExitCode` in the `BuildResult`, in the
`BuildCompleteEvent`. We do not include a `Throwable crash` in this scenario.

For the 'test'/'run' commands, the `BuildEventStreamer` was incorrectly
ignoring the error details specifically when the OOM error happened on
another thread and did not have a `Throwable crash` in its `BuildResult`.

A similar bug may exist for non-OOM catastrophic errors. Resolving this is left for a targeted followup.

PiperOrigin-RevId: 635654254
Change-Id: Ifdeb0f17463ca6d05d90077fbe6a2045c2a21ab2