Fixing pass by ref option_processor object Google C++ Linter warning
A linter warning appears with the pass by reference option_processor object.

The official Google C++ guideline say that the pass by ref/pointer options are to either make the reference parameter const if the parameter is not meant to be modified, or use a pointer instead of a non-const reference if the parameter is intended to be modified. Since "option_processor" is modified in the function definition, the change should be opted to pass by pointer.