[7.2.0] Add support for overlays in source.json (#22525)
This is a continuation of #22155 that adds the newly added
'remote_files' attribute for http_archive to the bzlmod functionality.

The end goal is to then update BCR to this new functionality to overlay
files rather than use patch files when providing MODULE/WORKSPACE/BUILD
files.

https://github.com/bazelbuild/bazel-central-registry/issues/1566 has a
good discussion of the rationale.

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>

Closes #22349.

PiperOrigin-RevId: 636682112
Change-Id: Ief070985598a7c0f427a98cd3daeb69a0984f7be

Commit
https://github.com/bazelbuild/bazel/commit/c4167e309a29384ee1cf827b4ecf4ff5b0210fc9

Co-authored-by: Farid Zakaria <fmzakari@google.com>