Do not record events announced after the build-completing event.
Doing so results in a race condition: last_message is set when there are no events that are announced but not posted. Progress events always announce the subsequent progress event so if there is a progress event after the build-completing event, last_message is never set.

The UI is updated on a separate thread and writing to the console results in a progress event so the above scenario is entirely possible.

RELNOTES: None.
PiperOrigin-RevId: 636810448
Change-Id: I86e1b9d37b338b3d574a6e27a7d72e81a6ecdffa