Automated rollback of commit bcbd956399c8322ede781ee5e2514de6171ca962.
*** Reason for rollback ***

https://github.com/bazelbuild/bazel/commit/4d6513b54fa07321115cac350553b29914d43a65 was innocent.

The issue was caused by a synchronization issue: even though BuildEventStreamer expects no events to be posted after BuildCompleteEvent, there is no guarantee that the UI thread does not do so and if it does, it breaks the logic that sets last_message.

This happens because progress events in response to console messages announce the next progress event in turn and last_message is set when there are no more announced but not posted events. Thus, any console write after BuildCompleteEvent but before the last event announced before BuildCompleteEvent is posted results in an event that is announced but never posted.

*** Original change description ***

Rollback of https://github.com/bazelbuild/bazel/commit/4d6513b54fa07321115cac350553b29914d43a65 due to user breakage reports.

PiperOrigin-RevId: 636821481
Change-Id: I9c2f8a51889fc03c054233beb31873d1dc63fd08