skyfocus: derive working set automatically.
This is a basic working set derivation that uses the source files found in the packages of the top level targets being built. Files in subdirectories are also included as long as they are in the same package.

Automatically derived working sets will start to become the default user journey, and user-defined working sets should only be used for testing or in debugging situations.

The exact semantics for the derived working sets are captured by the test cases.

Currently, the derivation will happen at the end of every build with `--experimental_enable_skyfocus=true`. Skyfocus will only run if the newly derived working set has expanded (new nodes), or if there's a new target to be focused.

Not in this CL, but: changes to the derivation on subsequent builds for already focused targets will be using signals from the version control system (i.e. Blaze will expand the working set with changed files automatically, with the minimal necessary reanalysis to build the missing nodes).

Example: `blaze build //a:b //c/d:e --experimental_enable_skyfocus` without `--experimental_working_set`, Blaze will infer the working set as all source files found in the Skyframe graph that belongs to the packages of the top level targets being built (`//a` and `//c/d`).
PiperOrigin-RevId: 637565764
Change-Id: I5b9726958833ff3f03697f01cf91ae80ae1547f4