Do not record events announced after the build-completing event, including by flushing stdout/stderr.
Doing so results in a race condition: `last_message` is set when there are no events that are announced but not posted. Progress events always announce the subsequent progress event so if there is a progress event after the build-completing event, `last_message` is never set. We prevent this from happening by ignoring announced/posted progress events after the build-completing event.

The UI is updated on a separate thread and writing to the console results in a progress event so the above scenario is entirely possible.

RELNOTES: Progress is no longer written to BEP after the build completing event is posted.
PiperOrigin-RevId: 638000909
Change-Id: Id30dbf756a5dbbd02ccf4281311c34040655f159