Write cc_common.link in Starlark
This code will replace cc_common.link. It is calling into cc_linking_helper, which is coming next.

Reorder the parameters. This made it easier to describe what the large number or parameters are actually doing in the linking process.

Add docs and default value for additional_outputs and variable_extensions. Those parameters were previously made public (not accidentally), but left undocumented.

PiperOrigin-RevId: 639020883
Change-Id: I13c1e424ae302448b883e8d3f009df5168776c76