Write create_linking_context_from_compilation_outputs in Starlark
This code will replace cc_common.create_linking_context_from_compilation_outputs. It is calling into cc_linking_helper, which is coming next.

Reorder the parameters. This made it easier to describe what the large number or parameters are actually doing in the linking process.

Add docs and default value for variable_extensions. This parameters was previously made public (not accidentally), but left undocumented.

PiperOrigin-RevId: 639027332
Change-Id: Ibc22b7227305514eb91d18c3a98772f99740bf04