Make it possible to determine whether a fileset symlink points into a tree artifact.
When a tree artifact is included in a Fileset, it is expanded and one symlink is created for each child file. This change stores the enclosing tree artifact in the `FilesetOutputSymlink` for those children. Note that this extra field does not change the shallow heap size of `FilesetOutputSymlink`.

Test coverage is added to the existing test for tree artifact inputs in `FilesetIntegrationTest`. An extra target is added to this test to cover the case of a nested fileset, which covers the change to `SkyframeFilesetManifestAction`.

PiperOrigin-RevId: 640195717
Change-Id: Ic275c25bdd92c7b87de78f9ba335ed9ee6dc561b