Split `SkyframeTests` into individual `java_test`s, improve `BUILD` granularity and test history tracking.
This halves the build time for a single change to a test file (9.6s [1] to 4.9s [2]) by not rebuilding `SkyframeTests_lib` all the time.

This target was added in 2015 (unknown commit) and has been growing since.

Also add a missing `java_test` for `MacOSXFsEventsDiffAwarenessTest.java`. It was excluded from `SkyframeTests_lib` and didn't have a corresponding test target, and all of its test methods was annotated with `@Ignore`. We could also delete it since it hasn't been running at all.

PiperOrigin-RevId: 640810575
Change-Id: I88a37beb64d5908ccd48099bd8a7146026153642