Refactor `FlagSetFunction` to depend on a `ProjectValue` directly, instead of depending on its own `BzlLoadValue`.
The dependency chain is now `FlagSetFunction` -> `ProjectFunction` -> `BzlLoadFunction`.

`FlagSetFunction` is kept distinct from `ProjectFunction` so as to not burden other reverse dependencies of `ProjectFunction` from the extra Skyframe lookup costs from `FlagSetFunction` (i.e. `ParsedFlagsFunction`). `ProjectFunction` is implemented as an indirection to `BzlLoadValue`, so there's a place for PROJECT.scl specific validations and getters in the future.

PiperOrigin-RevId: 641126914
Change-Id: I4e799395817c8d0f96a0af3e6c6520b29fafdf08