Report errors for all files with Error Prone diagnostics, not just the first one
This removes logic that was skipping Error Prone of compilations with errors. The intent was to skip analyzing compilations with javac errors, but it also meant that if Error Prone reported an error analyzing a file all subsequent files were skipped.

As of unknown commit, `ErrorProneAnalyzer` handles skipping analysis of compilations with errors (and distinguishing between javac and Error Prone errors), so it's safe for JavaBuilder to unconditionally run ErrorProne and let it decided whether or not to proceed.

This change also removes some bookkeeping about the number of 'flow' events javac processes and skips, which was used for testing. Since it no longer skips flow events, there isn't a good way to perform the assertion that was done previously. Since that logic was added we have better end to end integration tests for the Error Prone JavaBuilder integration, which reduces the risk of accidentally turning Error Prone off.

PiperOrigin-RevId: 641283846
Change-Id: I539f0713879835f031cb13a015b45dfa90e20c24