Fix toolchain type alias resolution with auto exec groups
The toolchain resolution for aliases in auto exec groups was broken because the
toolchain type labels were not being properly resolved. This was because the
`getAutomaticExecGroupLabels()` method was returning the exec group labels
instead of the toolchain type labels. This CL fixes the issue by
updating the code to return all the resolved toolchain type labels to include the actual labels of aliases as well.

This is supported without automatic execution groups as the added tests show.

PiperOrigin-RevId: 641990120
Change-Id: I17f555ef554ab1b8e18d61017513f35b4d39832c