Apply strategy policy to regex filters.
After this change the rules described in strategy policy proto file will be applied to action which strategies are described in `--regex_filter` flag.

Unfortunately we can't determine mnemonic of actions described in `--regex_filter`, so we need to apply the filter during the `getStrategies` method execution.

In cases when multiple filters are applicable for the action the behaviour is:
- If after sanitizing  the list of strategies is non-empty, we return this list.
- If after sanitizing  the list of strategies is empty, we fallback to default strategies for this mnemonic

PiperOrigin-RevId: 642557349
Change-Id: I3a70ae4177c36eeb074ca4e3dc57e0b8d6ee0cd5