JDK 21 update has broken the HeapOffsetHelper logic, fix it and make it more flexible to future brekages.
The most recent minor JDK changed the internals of the G1 garbage collector and
broke the workaround to JDK memory bean including internal G1 GC objects.  This
CL updates the logic to better handle this case with a regex which can be
updated without requiring a release. The current default value matches both
existing jdk.internal.vm.FillerArray and the new jdk.internal.vm.FillerEntries.
Until we get a better way (ideally a proper API from OpenJDK) we'll have to
keep updating to keep up with the internals.

PiperOrigin-RevId: 643081114
Change-Id: Ie0de433ddd41c1adaa6a10adf844a1b6dbb747af