Be more precise about where to read the URL rewriter config from
There is an [open issue](https://github.com/bazelbuild/bazel/issues/22104)
where bazel will occasionally fail at start up time because it
is unable to parse the downloader config. While I've yet to
create a reliable reproduction for this issue, the one time I've
triggered this error with a build with additional logging, it
seemed to be because the current working directory wasn't the
root of the workspace. While I'm not sure this will actually
fully resolve the problem (since I don't know what's causing it)
being clearer about where the config file is meant to be can
only be a Good Thing.