Implement mandatory `--scl_config` builds.
Supports projects that want to require developers to build with an authorized project-sanctioned set of build flags.

Gated by `--enforce_canonical_configs`.

See tests for exact `PROJECT.scl` syntax. This syntax is experimental and likely to change soon. Don't rely on it yet. For now we're just trying to get the core functionality together.

PiperOrigin-RevId: 644534201
Change-Id: I145b1bb1feca711977fbda2889ba6b159a8acb3c