Fix tests that deadlock with `--jobs=1`.
By default, `--jobs` is calculated from the number of host CPUs. In a remote execution environment, it may be 1. Certain rewinding tests use artificial synchronization to test race conditions and require multiple execution phase threads to guarantee progress.

Also mark `TopLevelActionLookupKeyWrapper` (the key type for `CompletionFunction`) as an `ExecutionPhaseSkyKey` so that it runs in the pool sized for `--jobs`.

PiperOrigin-RevId: 645247242
Change-Id: I83a957af04f2580d5d29ad7989d5befab80f1ce6