Consolidate special handling of executable path
In `CommandLines`, the very first argument of the first command line is always a path to an executable. As such, it should both be path mapped and turned into a "callable" path (e.g. `./tool.sh` instead of `tool.sh`). This wasn't the case for `SpawnAction`'s created via `ctx.actions.run(executable = <some string>)`.

This is fixed by moving the handling of the first argument into `CommandLines`.