[7.3.0] Push `makeExecutable` down into `AbstractFileWriteAction` subclasses (#22845)
Getting rid of the single boolean field on `AbstractFileWriteAction`
reduces padding on each subclass instance and in particular frees up a
4-byte field on `CppModuleMapAction`.

Also use a lambda to define `newDeterministicWriter` if possible for
improved readability.

This prepares for future changes that will add fields to
`CppModuleMapAction` to support path mapping.

Work towards #6526

Closes #22609.

PiperOrigin-RevId: 643340715
Change-Id: Id3af26049098e6dfa731f0e7a1be6709bea0d9f2

Closes #22749 
Closes #22750