Optimize Profiler.
Currently, we use `LinkedBlockingQueue` to queue up the trace event. But it uses `ReentrantLock` under the hood which means each time we start profiling a block, the `enqueue` method could block the thread - the time for blocking is included in the time span for the block itself which is confusing. The overhead adds ~20ms overhead to each profiled block when we have lots of trace event, e.g. with `--noslim_profile` and `--experimental_profile_additional_tasks=SKYFUNCTION`.

This CL optimizes this by using a lock-free queue, namely `ConcurrentLinkedQueue` and use best effort (explained in the code comment) to signal the consumer. Additionally, it minimizes operations when dealing with virtual lanes. This bring down the overhead to less than 1us.

PiperOrigin-RevId: 645430908
Change-Id: I68fab914dc0eac717260ea6cb43793c58bcddc0a