Refactor Package.Builder#fromOrFail variants
This is a cleanup that attempts to rationalize the various ways of validating that the current Starlark context is a BUILD thread, or a WORKSPACE thread, or a symbolic macro thread, or some combination of the above. It increases the total lines of code, but it avoids the awkward idiom of calling two different fromOrFail* methods to assert "not this context AND not that context either".

The semantics of some callables are slightly modified to prohibit usage in places where they probably wouldn't/couldn't be called anyway, e.g. vardef now disallows WORKSPACE usage, although it wasn't a top-level symbol in WORKSPACE anyway.

This is a follow-up to reviewer discussion from unknown commit.

PiperOrigin-RevId: 649080769
Change-Id: I3742b44f1d13d5cf81ef2e256d2a430dfa124813