Introduce a new `BuildPhase` enum to associate a command to Bazel's build phases.
This is a no-op change. The new enum is used to replace the boolean `builds` field, and all `builds()` calls are replaced by either `analyzes()` or `executes()`, whichever is more specific to the situation.

The new enum is hierarchical: `EXECUTES` implies `ANALYZES`, but `LOADS` does not imply `ANALYZES`. This is to reflect the fact that (for example) `EXECUTES` commands also analyze to create the configured target graph, while `LOADS` commands create the target graph or bzlmod graphs, and do not analyze. This will now allow us to distinguish between commands that only load BUILD/bzl files and commands that executes actions.

PiperOrigin-RevId: 649340932
Change-Id: I2a8319d738a5132943a91bd6691910d4be136d05