Test proto_common.compile
BEGIN_PUBLIC
Test proto_common.compile

Rewrite compile tests from BazelProtoCommonTest to Starlark. This is using rules_analysis for testing.

The tests are super fast (cca. 1s for all of the to run).

The tests work either with a redirect (calling native rule) or with actual implementation in the protobuf repository.
END_PUBLIC

PiperOrigin-RevId: 651748083
Change-Id: I40d5ab3c973e00243959a119920dfb868cf85d87