Automated rollback of commit 27a8ce409a05afb5e505e82d2c883e880296dc11.
*** Reason for rollback ***

The `runnable`s/`rejectionObserver`s can still be skipped if main skyframe-evaluator thread is interrupted and they are still waiting in the `ThreadPoolExecutor#workQueue`. So this change is not necessarily needed to fix b/349119929.

Post this rollback, another cl will be made to describe when runnable is going to be skipped. Also, we may choose to re-submit this change if not executing runnable becomes an issue.

See discussion in unknown commit for more details.

*** Original change description ***

Add the `executeWithFallback()` method in `AbstractQueueVisitor` which takes in a second `RejectedExecutionObserver` parameter.

The fallback `RejectedExecutionObserver#onRejectedExecution()` is executed when the `originalRunnable` is blocked from execution due to exception caught. This is necessary when the `originalRunnable` changes synchronization state. So failure to execute the `originalRunnable` leads to deadlock.

PiperOrigin-RevId: 652981345
Change-Id: I2d7c50664881f6d46d65b30976770806692a8c6d