virtual-thread: use consistent separator in thread name
As the profile visualization is often made up of mutliple horizontal
lanes stacked on top of each others, each lane representing a unique
thread, using Virtual Thread would create more unique lanes in Bazel
profile visualization and make it harder to see.  Most of the time,
these virtual thread are not reused and thus, would only contain a
single column of parent-child spans. It would be easier to consume if
we can group up the virtual threads of the same category under a
single lane.

Rename some of the existing virtual thread factories to use hyphen as the
separator in the threads' name. This would help downstream toolings to
identify the thread's category easier.