Test action interaction with platform
Add a test to show how no-remote-exec exec prop work w. test action

The most recent test added demonstrated that if `no-remote-exec` is
added to a platform and action were to run on that platform, then that
action will be executed locally instead of remotely. The intended use
case for this was to setup a local "fallback" exec platform that
actions could run on if none of the registered exec platform could be
used. However, this does not hold with test action.

Each test target in Bazel is composed of 2 types of actions:

a. One to build the test executable binary
b. One to run the executable binary to obtain test logs and result

Actions in group (a) are executed on the matched `exec` platform and
produce the artifact intended for the `target` platform. This artifact
is then used in (b), which means that actions in (b) must be executed
in `target` platform, defined by `--platforms` flag.

Provide a test to demonstrate that currently, there is no way to tell
Bazel to execute this test action (b) in the local fallback platform.

Closes #23037.

PiperOrigin-RevId: 653921346
Change-Id: I354686d2bacc6fce32e10abf8f3df98ac8e4ca07