Do not wrap object files in `--start-lib/--end-lib` for nodeps libraries
When linking a nodeps dynamic library that contains object files
produced from tree artifact sources, the library may not contain any
references to the symbols in those files. In this case, when wrapping
the files in `--start-lib/--end-lib`, they may end up being dropped by
the linker, which results in missing symbols compared to a build with
`--dynamic_mode=off`.

This is fixed by not wrapping object files into `--start-lib/--end-lib`
for the nodeps dynamic library link action.