Also check if computed default inputs are configurable in isConfigurable
Fixes a crashing bug that occurs when there's a computed default
depending on a configurable attribute along with an implicit output
function. There does not need to be a dependency on the computed
attribute in the implicit output function.

When calculating the set of attribute values, configurable attributes
should be omitted. But we must also make sure computed defaults with
configurable inputs are omitted. It's safer all around if these also
count as "configurable" since they should be similarly "restricted".

PiperOrigin-RevId: 655547739
Change-Id: I0ca79e70bf2bddf003df20a57e5b858c5ea8e02a