Clean the starlark transition cache more eagerly.
This causes starlark transitions to be re-evaluated on every build, but ensures that they are correct when the underlying sources change.

A better option would be to use skyframe's existing cache mechanism to clear these only when the related FileStateKeys change, but that's not possible to detect in the current mechanism.

Fixes #23097.

Closes #23100.

PiperOrigin-RevId: 656096828
Change-Id: I1204ebed09cd9e155b4cfa430330d68ea55b8a99