Properly handle selects with None values for label-valued attrs in native.existing_rule and native.existing_rules
Label-valued attributes have a default value of null in Java. The BuildType.Selector
constructor casts None values of select branches to that default - i.e. to null; but
we must transform that null back to None when reconstructing a SelectorValue in
native.existing_rule/s for use in Starlark - or we would miss a branch.

And take the opportunity to ensure we never attempt to construct an empty
SelectorValue or SelectorList - which is a fatal error.

PiperOrigin-RevId: 656122706
Change-Id: Id236af711ba7f7c515a1b238e4be9d7d63574173