Update `--compile_one_dependency` heuristics for header-only libraries
Previously, if a header was part of a header_only library :a, then
`--compile_one_dependency` would prefer to choose :b which depends on `:a`, rather
than `:a` itself.
After this change, it will generally prefer `:a`, unless parse_headers is turned
off specifically in this package (in which case it is likely that a.h can only
be understood in the context of `b.cc`).

RELNOTES: `--compile_one_dependency` selects header-only `cc_library`s in more cases
PiperOrigin-RevId: 656460540
Change-Id: Iba70dddd1610a78d71c05274d77d479c84efade7