[7.3.0] Ignore transitive cppmap files from dotd files (#23122)
When clang generates dotd files when using `-fmodule-map-file` any
`extern module` directives in the modulemap are included in the dotd
file if they exist. The result of this was that with sandboxing disabled
the dotd file included transitive cppmap files that weren't in its input
set, resulting in build failures. This change excludes those instead
since they're not required as evidence by the fact that with sandboxing
enabled they are not part of the input set.

Fixes https://github.com/bazelbuild/bazel/issues/21592

Closes #21832.

PiperOrigin-RevId: 656382428
Change-Id: I4bc9802884ce1bc66ceda65a602db8dffbd1d9ea

Commit
https://github.com/bazelbuild/bazel/commit/ad53147cc4f4d47c064845ec9612876161ff139e

Co-authored-by: Keith Smiley <keithbsmiley@gmail.com>