Make Window launchers relocatable
When the Windows launcher binary is copied to become the `executable` of another target, it can't expect to find the main file right next to it. Instead, look it up via `Rlocation`, which doesn't require colocation to find the file. This helps with a common pattern in wrapper rules where `ctx.symlink` is used to work around the limitation that the `executable` of a target has to be produced by that target.

For the Python rules, for which the source of truth is not the Bazel repo anymore, the new launch key is checked for in a backwards compatible way.

Closes #23076.

PiperOrigin-RevId: 657116461
Change-Id: I3f966b48e0812668cfa7bd394f2eaf23d66889b6