Add an explicit error when trying to expand a `TreeArtifact` outside of inputs list in `Args.add_{all,joined}`
This is the analogue of b7590a03b48b353b72415a1896d014130740b535, which did this for `Fileset`.

Previously, when users forgot to add a `TreeArtifact` (aka a "directory") to the inputs of an action, `Args` would silently expand the contents to the empty set and thus (usually) not add anything to the command line. This made the mistake difficult to track down.