[7.3.0] Add flag --incompatible_use_plus_in_repo_names (#23141)
Defaults to false. When set to true, we use `+` instead of `~` as the
separator in canonical repo names.

Some more subtle changes include:
- We now officially say that the "build" part of version strings (the
part that begins with a plus) is ignored and stripped.
- When the flag is set to true, we effectively increase the lockfile
version by 1 (see code comment in BazelLockFileModule).
- When true, we no longer insert a `_main` in front of names of repos
generated by module extensions hosted in the main repo. (`~abc` as a
name was problematic, but `+abc` is not.)
- When true, we no longer insert a `v` in front of numerical versions in
canonical repo names. (`my_mod~1.1` could be a Windows short path, but
`my_mod+1.1` cannot.)

Work towards https://github.com/bazelbuild/bazel/issues/22865.

Closes #23103.

PiperOrigin-RevId: 657202616
Change-Id: I015b2a04a823b1d951015a1b2e1b99b154dcc5a2

Commit
https://github.com/bazelbuild/bazel/commit/2f67e57899425a7adad0ba03d524d6287902b11f

Co-authored-by: Xdng Yng <wyverald@gmail.com>